move pickledb extensions to util
