+++ otherdave/util/__init__.py
@@ -2,4 +2,5 @@
 
 from .madlib import *
 from .dlog import *
+from .pickledbrg import *
 from .triggers import *
